+++ tests/src/Unit/Event/AlterProductDataEventTest.php
@@ -5,6 +5,7 @@ namespace Drupal\Tests\commerce_trustedshops\Unit\Event;
 use Drupal\commerce_order\Entity\OrderItemInterface;
 use Drupal\commerce_trustedshops\Event\AlterProductDataEvent;
 use Drupal\Tests\UnitTestCase;
+use Prophecy\Prophet;
 
 /**
  * @coversDefaultClass \Drupal\commerce_trustedshops\Event\AlterProductDataEvent
@@ -14,13 +15,37 @@ use Drupal\Tests\UnitTestCase;
 class AlterProductDataEventTest extends UnitTestCase {
 
   /**
+   * The prophecy object.
+   *
+   * @var \Prophecy\Prophet
+   */
+  private $prophet;
+
+  /**
+   * {@inheritdoc}
+   */
+  public function setUp(): void {
+    parent::setUp();
+    $this->prophet = new Prophet();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function tearDown(): void {
+    parent::tearDown();
+
+    $this->prophet->checkPredictions();
+  }
+
+  /**
    * @covers ::getProductData
    * @covers ::setProductData
    * @covers ::getOrderItem
    */
   public function testEvent() {
     /** @var \Drupal\commerce_order\Entity\OrderItemInterface $orderItem */
-    $orderItem = $this->prophesize(OrderItemInterface::class)->reveal();
+    $orderItem = $this->prophet->prophesize(OrderItemInterface::class)->reveal();
 
     $event = new AlterProductDataEvent([
       'name' => 'Initial product',