+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add official support of drupal 10.0
+
+### Changed
+- re-enable PHPUnit Symfony Deprecation notice
 
 ## [2.0.0] - 2023-01-27
 ### Fixed