+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - fix PHPUnit deprecated prophecy integration
+- fix deprecated uncalling accessCheck relying on entity queries to check access
 
 ## [2.0.0] - 2023-01-27
 ### Fixed